+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - PR #1165 updated remaining algorithms to be NetworkX compatible
 - PR #1176 Update ci/local/README.md
 - PR #1184 BLD getting latest tags
+- PR #1222 Added min CUDA version check to MG Louvain
 - PR #1217 NetworkX Transition doc
 - PR #1223 Update mnmg docs
 
@@ -47,7 +48,6 @@
 - PR #1180 BLD Adopt RAFT model for cuhornet dependency
 - PR #1181 Fix notebook error handling in CI
 - PR #1199 BUG segfault in python test suite
-- PR #1186 BLD Installing raft headers under cugraph 
 - PR #1186 BLD Installing raft headers under cugraph
 - PR #1192 Fix benchmark notes and documentation issues in graph.py
 - PR #1196 Move subcomms init outside of individual algorithm functions